// src/schemas/BatteryStatus.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Battery status */
export type BatteryStatus = {
  /** Standard ROS header */
  header: Header;

  /** Battery discharge percent */
  battery_discharge_percent: number;

  /** Battery current (A) */
  battery_current: number;

  /** Battery voltage (V) */
  battery_voltage: number;
};