// src/schemas/GeneralVehicleStatus.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** General vehicle status */
export type GeneralVehicleStatus = {
  /** Standard ROS header */
  header: Header;

  signal12_switch: boolean;
  request_zero_throttle: boolean;
  request_brake: boolean;
  signal_brights_on: boolean;
  signal_right_turn: boolean;
  signal_left_turn: boolean;
  signal_horn: boolean;
  signal_hazard_lights: boolean;
  signal_fog_lights: boolean;
  signal_direction_reverse: boolean;
  global_man_sig_mag_brake: boolean;
  global_curtis_sig_mag_brake: boolean;
  signal_brake_switch: boolean;
  signal_seat_switch: boolean;
  e_stop_status: boolean;
  selected_application: number;
  selected_op_mode: number;
  active_op_mode: number;
  button_blue: boolean;
  button_yellow: boolean;
  button_green: boolean;
};