// src/schemas/RawSignalSteeringForce.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Raw steering force signal */
export type RawSignalSteeringForce = {
  /** Standard ROS header */
  header: Header;

  uint32_signal_steering_force_b: number;
  uint32_signal_steering_force_a: number;
};