// src/schemas/RawSignalSteeringPosition.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Raw steering position signal */
export type RawSignalSteeringPosition = {
  /** Standard ROS header */
  header: Header;

  uint16_encoder_raw_value_b: number;
  uint16_encoder_raw_value_a: number;
};